requirements.txt
streamlit
transformers
torch
PyMuPDF
bert-extractive-summarizer
scikit-learn
numpy
